requirements.txt
flask
flask_sqlalchemy
flask_login
flask_migrate
psycopg2-binary
python-dotenv
werkzeug
reportlab
qrcode
Pillow
flask-wtf
